LayoutTests/imported/w3c/web-platform-tests/fetch/range/general.any.worker-expected.txt

CONSOLE MESSAGE: Preflight response is not successful. Status code: 404
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 500
CONSOLE MESSAGE: Preflight response is not successful. Status code: 404

PASS Range header setting allowed for guard type: none
PASS Range header setting allowed for guard type: response
PASS Range header setting allowed for guard type: request
PASS Privileged header not allowed for guard type: request-no-cors
PASS Fetch with range header will be sent with Accept-Encoding: identity
PASS Cross Origin Fetch with non safe range header
PASS Cross Origin Fetch with safe range header
